+++ comparison_to_other_codes/laminar_150lmm/laminar_150lmm_monochromator_sweep.py
@@ -72,6 +72,7 @@ runner = rp.BatchSimulationRunner(
     on_error="fail_fast",
     max_workers='auto',
     resume=False,
+    backend="numba",
 )
 
 grating.plot_profile(profile_plot_path)